attempt to help :)
After ~1 hour of attempts, I couldn't find a solution to avoid `undefined` into the return type. I've modified this code to don't throw an exception when a required input is read, but it still brings `undefined` into the return type.
I don't know how result of a query can be returned before the query, but it's just details of the implementation. Maybe there is some "dummy result" that is returned before the query is executed, I don't know, but then you could use that as a replacement for undefined.

This commit is not supposed to be a PR, it is just part of the conversation.